requirements.txt
certifi==2024.2.2
charset-normalizer==2.1.0
docker==6.1.3
fasteners==0.17.3
gitdb==4.0.9
GitPython==3.1.42
idna==3.3
inotify-simple==1.3.5
packaging==24.0
python-decouple==3.6
PyYAML==6.0
requests==2.31.0
smmap==5.0.0
urllib3==2.2.1
websocket-client==1.3.3
python-dotenv==1.0.1
